requirements.txt
# Requirements for the project

# For data processing & ML
pandas
numpy
scikit-learn
torch
transformers
transformers[torch]
accelerate


# For NLP & text processing
spacy
presidio-analyzer
presidio-anonymizer


# For database
SQLAlchemy
pymysql


# For deployment & ui
huggingface_hub
streamlit
plotly-express


#Utilities
python-dotenv